Call cpufreq_update_policy on DVFS events and force SLEEP_FREQ on suspend.
Calls cpufreq_update_policy on DVFS lock/unlock events to ensure that DVFS
locks take effect with the governor's preferred frequency (if higher than
the lock frequency).  Previously the lock would set the frequency to the
target level and expect the governor to reevaluate on a timer, which
ondemand does, but powersave and performance do not.

Forces the cpufreq policy, when suspending, to include SLEEP_FREQ, so as to
prevent the device from crashing due to use of a "wrong" frequency.  Commit
is based on [1], but includes additional locking to ensure policy isn't
overwritten just prior to calling cpufreq_driver_target.  Also ensures that
the user-specified cpufreq policy is reevaluated on resume, which is not
done in the stock version.

[1] https://github.com/nubecoder/nubernel-2.6.35/commit/6255b550af8c4f9d342ac4806e382dccfb40f4b6

Change-Id: Ia50f1c9d1365ddfbcf4543f8acb4d00984284388